refactor : timeline -> mypage로 디렉토리명 변경
