+++ playwithme/src/main/java/com/idea5/playwithme/event/domain/Event.java
@@ -1,15 +1,10 @@
 package com.idea5.playwithme.event.domain;
 
-import com.idea5.playwithme.comment.domain.Comment;
-import com.idea5.playwithme.timeline.domain.Timeline;
 import lombok.*;
-import com.idea5.playwithme.board.domain.Board;
 import org.springframework.data.annotation.CreatedDate;
 
 import javax.persistence.*;
 import java.time.LocalDateTime;
-import java.util.ArrayList;
-import java.util.List;
 
 @Entity
 @Getter
