+++ playwithme/src/main/java/com/idea5/playwithme/member/domain/Member.java
@@ -2,7 +2,7 @@ package com.idea5.playwithme.member.domain;
 
 import com.idea5.playwithme.article.domain.Article;
 import com.idea5.playwithme.comment.domain.Comment;
-import com.idea5.playwithme.timeline.domain.Timeline;
+import com.idea5.playwithme.mypage.domain.Timeline;
 import com.idea5.playwithme.together.domain.Together;
 import lombok.*;
 import org.springframework.data.annotation.CreatedDate;
