+++ playwithme/src/main/java/com/idea5/playwithme/mypage/dto/TimelineRequestDto.java
@@ -1,10 +1,8 @@
-package com.idea5.playwithme.timeline.dto;
+package com.idea5.playwithme.mypage.dto;
 
-import com.idea5.playwithme.article.domain.Article;
-import com.idea5.playwithme.comment.domain.Comment;
 import com.idea5.playwithme.event.domain.Event;
 import com.idea5.playwithme.member.domain.Member;
-import com.idea5.playwithme.timeline.domain.Timeline;
+import com.idea5.playwithme.mypage.domain.Timeline;
 import com.idea5.playwithme.together.domain.Together;
 import lombok.Getter;
 import lombok.Setter;
