+++ playwithme/src/main/java/com/idea5/playwithme/mypage/exception/DataNotFoundException.java
@@ -1,4 +1,4 @@
-package com.idea5.playwithme.timeline.exception;
+package com.idea5.playwithme.mypage.exception;
 
 public class DataNotFoundException extends RuntimeException {
     public DataNotFoundException(String question_not_found) {
