+++ playwithme/src/main/java/com/idea5/playwithme/mypage/repository/TimelineRepository.java
@@ -1,6 +1,6 @@
-package com.idea5.playwithme.timeline.repository;
+package com.idea5.playwithme.mypage.repository;
 
-import com.idea5.playwithme.timeline.domain.Timeline;
+import com.idea5.playwithme.mypage.domain.Timeline;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
