+++ playwithme/src/main/java/com/idea5/playwithme/mypage/service/TimelineService.java
@@ -1,14 +1,14 @@
-package com.idea5.playwithme.timeline.service;
+package com.idea5.playwithme.mypage.service;
 
 import com.idea5.playwithme.article.domain.Article;
 import com.idea5.playwithme.article.repository.ArticleRepository;
 import com.idea5.playwithme.event.domain.Event;
 import com.idea5.playwithme.member.domain.Member;
 import com.idea5.playwithme.member.repository.MemberRepository;
-import com.idea5.playwithme.timeline.domain.Timeline;
-import com.idea5.playwithme.timeline.dto.TimelineRequestDto;
-import com.idea5.playwithme.timeline.exception.DataNotFoundException;
-import com.idea5.playwithme.timeline.repository.TimelineRepository;
+import com.idea5.playwithme.mypage.domain.Timeline;
+import com.idea5.playwithme.mypage.dto.TimelineRequestDto;
+import com.idea5.playwithme.mypage.exception.DataNotFoundException;
+import com.idea5.playwithme.mypage.repository.TimelineRepository;
 import com.idea5.playwithme.together.domain.Together;
 import com.idea5.playwithme.together.repository.TogetherRepository;
 import lombok.RequiredArgsConstructor;
