+++ playwithme/src/main/java/com/idea5/playwithme/together/service/TogetherService.java
@@ -5,7 +5,7 @@ import com.idea5.playwithme.article.domain.Article;
 import com.idea5.playwithme.article.repository.ArticleRepository;
 import com.idea5.playwithme.member.domain.Member;
 import com.idea5.playwithme.member.repository.MemberRepository;
-import com.idea5.playwithme.timeline.exception.DataNotFoundException;
+import com.idea5.playwithme.mypage.exception.DataNotFoundException;
 import com.idea5.playwithme.together.domain.Together;
 import com.idea5.playwithme.together.repository.TogetherRepository;
 import lombok.RequiredArgsConstructor;
